# ARROW-10795: [Rust] Optimize specialization for datatypes