@Dandandan 2020-12-03
That makes also sense to me, the code containing conditions like this for two different datatypes is a sign that the current design is too complicated.