@Dandandan 2020-12-03
I did also some new profiling using callgrind. It seems the compiler is still not always able to fully specialize even for the `if matches!(T::DATA_TYPE, DataType::Boolean) ` check, though it is still much faster / needs way less instructions. So there is also still a tiny perf improvement possible.